requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Authentication packages
django-allauth>=0.54.0
django-allauth-ui>=0.1.3
dj-rest-auth>=4.0.1
djangorestframework-simplejwt>=5.3.0
oauthlib>=3.2.2
requests-oauthlib>=1.3.1
python-jose>=3.3.0
cryptography>=41.0.0

# Social authentication providers
# Google
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Facebook
facebook-sdk>=3.1.0

# Apple
PyJWT>=2.8.0

# Admin panel
django-jazzmin>=2.6.0

# Other dependencies
Pillow>=9.5.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.6
requests>=2.31.0
qrcode[PIL]
markdown>=3.4.0
telegram-webapp-auth==0.1.0

# Firebase
firebase-admin>=6.2.0
